Openfire plugin should compile
After a code change in a dependency, the Openfire plugin failed to compile.
The upstream code change introduced an assumption that the XMPP component
will act as an external component (XEP-0114). Although valid in many
deployments, this is not the case when the component is executed inside
an Openfire plugin.

As the component in an Openfire plugin is not being executed as an external
component, most of the constructor argument values that are new defined in
the upstream API (which broke compilation in the first place) won't be used
at runtime. A best-effort approach has been made to populate those arguments
with 'correct' values none-the-less.