Initialize OSGI context
Jitsi Videobridge makes use of an OSGI context that needs to be initialized
by the Openfire plugin.

This commit overrides the default OSGI bundle that ships with Jitsi, in order
to disable the REST API that's normally started. By default, Openfire would
not allow for an arbritary new port to be opened for web traffic (if the
existing implementation of that bundle would not thow ClassNotFoundExceptions
as it does now). There is room for improvement here, though.