Exceptions should not be completely ignored
This is a fix for https://github.com/jitsi/jitsi-videobridge/issues/342

When an exception occurs, it should not be ignored completely. If handling is not needed, the exception should be logged (if only at the lowest of log levels). This greatly reduces confusion while debugging a situation where a bug occurs caused by an exception that the original developer believed was "impossible".